requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
asyncio==3.4.3
attrs==25.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
setuptools==76.0.0
Werkzeug==3.1.3
yarl==1.18.3
